move prepare phase to an extra .py file
